Renamed models dir to types
